requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
pillow==12.1.1
sqlparse==0.5.5
tzdata==2025.3
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
